+++ Kernel/System/DynamicField.pm
@@ -191,13 +191,12 @@ sub DynamicFieldAdd {
     # sql
     return if !$Self->{DBObject}->Do(
         SQL =>
-            'INSERT INTO dynamic_field (internal_field, name, label, field_Order, field_type, object_type,'
-            .
-            'config, valid_id, create_time, create_by, change_time, change_by)' .
+            'INSERT INTO dynamic_field (internal_field, name, label, field_Order, field_type, object_type,' .
+            ' config, valid_id, create_time, create_by, change_time, change_by)' .
             ' VALUES (?, ?, ?, ?, ?, ?, ?, ?, current_timestamp, ?, current_timestamp, ?)',
         Bind => [
             \$InternalField, \$Param{Name}, \$Param{Label}, \$Param{FieldOrder}, \$Param{FieldType},
-            \$Param{ObjectType}, \$Config, \$Param{ValidID}, \$Param{UserID}, \$Param{UserID},
+            \$Param{ObjectType}, \$Config, \$Param{ValidID}, \$Param{UserID}, \$Param{UserID}
         ],
     );