+++ Kernel/System/DynamicField/Backend/Date.pm
@@ -855,13 +855,6 @@ sub ObjectMatch {
     return 0;
 }
 
-sub AJAXPossibleValuesGet {
-    my ( $Self, %Param ) = @_;
-
-    # not supported
-    return;
-}
-
 sub HistoricalValuesGet {
     my ( $Self, %Param ) = @_;