+++ Kernel/System/DynamicField/Backend/DateTime.pm
@@ -817,13 +817,6 @@ sub ObjectMatch {
     return 0;
 }
 
-sub AJAXPossibleValuesGet {
-    my ( $Self, %Param ) = @_;
-
-    # not supported
-    return;
-}
-
 sub HistoricalValuesGet {
     my ( $Self, %Param ) = @_;